Implement GitHub user loading from API
This commit also introduces Hilt/Dagger DI.
GitHub API token is supplied via BuildConfig and is set in local.properties.
It also contains a fix for compilation issue in when projects
use buildSrc and it causes a version conflict of com.squareup.javapoet
transitive dependency and it needs to be explicitly set to the version that works.